requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

ipython==8.32.0
numpy==2.2.1
ollama==0.4.5
opencv-python==4.11.0.86
pdf2image==1.17.0
pillow==11.1.0
pytesseract==0.3.13
requests==2.32.3
scikit-image==0.25.2
